+++ tests/Unit/Order/Service/OrdereServiceTest.php
@@ -7,16 +7,16 @@
 
 declare(strict_types=1);
 
-namespace FreshAdvance\Invoice\Tests\Unit\Service;
+namespace FreshAdvance\Invoice\Tests\Unit\Order\Service;
 
 use FreshAdvance\Invoice\DataType\InvoiceDataInterface;
+use FreshAdvance\Invoice\Order\Service\OrderService;
+use FreshAdvance\Invoice\Order\Service\OrderServiceInterface;
 use FreshAdvance\Invoice\Repository\OrderRepositoryInterface;
-use FreshAdvance\Invoice\Service\OrderService;
-use FreshAdvance\Invoice\Service\OrderServiceInterface;
 use OxidEsales\Eshop\Application\Model\Order;
 use PHPUnit\Framework\TestCase;
 
-/** @covers \FreshAdvance\Invoice\Service\OrderService */
+/** @covers \FreshAdvance\Invoice\Order\Service\OrderService */
 class OrdereServiceTest extends TestCase
 {
     public function testsomthing(): void